echoflow-poc/requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
google-generativeai>=0.5.0
